Clean up frontend
